+++ apx_sdk/CHANGELOG.md
@@ -9,6 +9,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Added
 
 - Added detailed error description to `FetchError::InvalidProof` variant.
+- Added `send_media` function.
 
 ### Changed